[serve] restructure deployment scheduler tests
* Move some of the tests in `test_deployment_scheduler` to `unit/test_deployment_scheduler.py` and mock out cluster node info cache because they don't need to start ray.
* Add some methods to `MockClusterNodeInfoCache`.
